Move new TagSelectList component to separate module
