+++ src/components/organize/TagManager/components/TagSelect/index.tsx
@@ -1,0 +1,98 @@
+/* eslint-disable jsx-a11y/no-autofocus */
+import { useAutocomplete } from '@material-ui/lab';
+import { useIntl } from 'react-intl';
+import { useState } from 'react';
+import { Box, TextField } from '@material-ui/core';
+
+import { groupTags } from '../../utils';
+import TagDialog from '../TagDialog';
+import TagSelectList from './TagSelectList';
+import { EditTag, NewTag } from '../../types';
+import { ZetkinTag, ZetkinTagGroup } from 'types/zetkin';
+
+const TagSelect: React.FunctionComponent<{
+  disableEditTags?: boolean;
+  disabledTags: ZetkinTag[];
+  groups: ZetkinTagGroup[];
+  onCreateTag: (tag: NewTag) => void;
+  onEditTag: (tag: EditTag) => void;
+  onSelect: (tag: ZetkinTag) => void;
+  tags: ZetkinTag[];
+}> = ({
+  disableEditTags,
+  disabledTags,
+  groups,
+  onCreateTag,
+  onEditTag,
+  onSelect,
+  tags,
+}) => {
+  const intl = useIntl();
+
+  const [tagToEdit, setTagToEdit] = useState<
+    ZetkinTag | Pick<ZetkinTag, 'title'> | undefined
+  >(undefined);
+
+  const {
+    inputValue,
+    getInputProps,
+    getListboxProps,
+    getRootProps,
+    groupedOptions,
+  } = useAutocomplete({
+    getOptionLabel: (option) => option.title,
+    open: true,
+    options: tags,
+  });
+
+  const groupedFilteredTags = groupTags(
+    groupedOptions,
+    intl.formatMessage({
+      id: 'misc.tags.tagManager.ungroupedHeader',
+    })
+  );
+
+  return (
+    <Box {...getRootProps()}>
+      <TextField
+        {...getInputProps()}
+        autoFocus
+        fullWidth
+        inputProps={{
+          'data-testid': 'TagManager-TagSelect-searchField',
+        }}
+        placeholder={intl.formatMessage({
+          id: 'misc.tags.tagManager.addTag',
+        })}
+        variant="outlined"
+      />
+      {/* Options */}
+      <TagSelectList
+        disabledTags={disabledTags}
+        disableEditTags={!!disableEditTags}
+        groupedTags={groupedFilteredTags}
+        inputValue={inputValue}
+        listProps={getListboxProps()}
+        onEdit={(tag) => setTagToEdit(tag)}
+        onSelect={onSelect}
+      />
+      <TagDialog
+        groups={groups}
+        onClose={() => setTagToEdit(undefined)}
+        onSubmit={(tag) => {
+          if ('id' in tag) {
+            // If existing tag
+            onEditTag(tag);
+          } else {
+            // If new tag
+            onCreateTag(tag);
+          }
+        }}
+        open={Boolean(tagToEdit)}
+        tag={tagToEdit}
+      />
+    </Box>
+  );
+};
+
+export default TagSelect;
